ref: Flatten TSR adapter code
No need for an impl file here, it made sense in Next.js for the
agnostic adapter to import both app & pages router implementations.
