--- packages/nuqs/src/adapters/tanstack-router/index.ts
@@ -1,4 +1,0 @@
-import { createAdapterProvider } from '../lib/context'
-import { useNuqsTanstackRouterAdapter } from './impl'
-
-export const NuqsAdapter = createAdapterProvider(useNuqsTanstackRouterAdapter)
